+++ storage_drivers/storage_drivers.go
@@ -16,7 +16,7 @@ import (
 const ConfigVersion = 1
 
 // DriverVersion is the actual release version number
-const DriverVersion = "17.07.0"
+const DriverVersion = "17.07.1"
 
 // FullDriverVersion is the DriverVersion as well as any pre-release tags
 var FullDriverVersion = DriverVersion